update package.json to include luxon package
