+++ package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack"
   },
   "keywords": [],
   "author": "",
@@ -325,6 +326,7 @@
     "longest-streak": "^2.0.4",
     "lowercase-keys": "^2.0.0",
     "lru-cache": "^7.18.3",
+    "luxon": "^3.3.0",
     "make-dir": "^3.1.0",
     "make-fetch-happen": "^11.1.1",
     "map-age-cleaner": "^0.1.3",
@@ -498,12 +500,12 @@
     "specificity": "^0.4.1",
     "sprintf-js": "^1.0.3",
     "ssri": "^10.0.4",
+    "string_decoder": "^1.3.0",
     "string-width": "^5.1.2",
     "string-width-cjs": "^4.2.3",
     "string.prototype.trim": "^1.2.7",
     "string.prototype.trimend": "^1.0.6",
     "string.prototype.trimstart": "^1.0.6",
-    "string_decoder": "^1.3.0",
     "strip-ansi": "^6.0.1",
     "strip-ansi-cjs": "^6.0.1",
     "strip-bom": "^3.0.0",
@@ -566,6 +568,8 @@
     "w3c-xmlserializer": "^4.0.0",
     "wcwidth": "^1.0.1",
     "webidl-conversions": "^7.0.0",
+    "webpack": "^5.88.1",
+    "webpack-cli": "^5.1.4",
     "whatwg-encoding": "^2.0.0",
     "whatwg-mimetype": "^3.0.0",
     "whatwg-url": "^11.0.0",
