+++ public/index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Books Store</title>
-  <link rel="stylesheet" href="index.css">
+  <link rel="stylesheet" href="../src/index.css">
 </head>
 <body>
   <header>
@@ -60,6 +60,6 @@
   </div>
 
 
-  <script type="module"   src="index.js"> </script>
+  <script src="../dist/main.js"> </script>
 </body>
 </html>
